Migrating from XCode to Swift Package
