+++ clients/swift/SwiftLockbookCore/Sources/SwiftLockbookCore/Models/Error.swift
