+++ clients/swift/SwiftLockbookCore/Sources/SwiftLockbookCore/Models/File.swift
