+++ clients/swift/SwiftLockbookCore/Sources/SwiftLockbookCore/Models/FileMetadata.swift
