+++ clients/swift/SwiftLockbookCore/Tests/SwiftLockbookCoreTests/Resources/accountString.txt
