+++ clients/swift/SwiftLockbookCore/Tests/SwiftLockbookCoreTests/Resources/workResult.json
